requirements.txt
bottle==0.12.13
Flask==1.0.2
line-bot-sdk==1.8.0
requests==2.19.1
requests-oauthlib==1.0.0
urllib3==1.23